Improve maze display and refactor
Now the maze display actually has squares instead of stretched
monstrosities. Furthermore, it is now centered on the screen. I then
abstracted it away. It should be much faster now: all the rects to draw
are generted, and then they are all drawn at once through a single call.
It can also handle non-screen size drawing areas now. And you can
change the size of the borders with a single constant. This took way too
long, but it looks pretty good.